+++ app/src/main/java/com/nextcloud/talk/chat/ChatActivity.kt
@@ -195,7 +195,6 @@ import com.nextcloud.talk.utils.Mimetype
 import com.nextcloud.talk.utils.NotificationUtils
 import com.nextcloud.talk.utils.ParticipantPermissions
 import com.nextcloud.talk.utils.SpreedFeatures
-import com.nextcloud.talk.utils.UriUtils
 import com.nextcloud.talk.utils.VibrationUtils
 import com.nextcloud.talk.utils.bundle.BundleKeys
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_CALL_VOICE_ONLY
@@ -209,7 +208,6 @@ import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_ROOM_ID
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_ROOM_TOKEN
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_START_CALL_AFTER_ROOM_SWITCH
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_SWITCH_TO_ROOM
-import com.nextcloud.talk.utils.database.user.CurrentUserProviderNew
 import com.nextcloud.talk.utils.permissions.PlatformPermissionUtil
 import com.nextcloud.talk.utils.rx.DisposableSet
 import com.nextcloud.talk.utils.singletons.ApplicationWideCurrentRoomHolder
@@ -266,9 +264,6 @@ class ChatActivity :
 
     @Inject
     lateinit var ncApi: NcApi
-
-    @Inject
-    lateinit var currentUserProvider: CurrentUserProviderNew
 
     @Inject
     lateinit var permissionUtil: PlatformPermissionUtil
@@ -3298,34 +3293,6 @@ class ChatActivity :
                 )
             )
         )
-    }
-
-    override fun startActivity(intent: Intent) {
-        val user = currentUserProvider.currentUser.blockingGet()
-        if (intent.data != null && TextUtils.equals(intent.action, Intent.ACTION_VIEW)) {
-            val uri = intent.data.toString()
-            if (uri.startsWith(user.baseUrl!!)) {
-                if (UriUtils.isInstanceInternalFileShareUrl(user.baseUrl!!, uri)) {
-                    // https://cloud.nextcloud.com/f/41
-                    val fileViewerUtils = FileViewerUtils(applicationContext, user)
-                    fileViewerUtils.openFileInFilesApp(uri, UriUtils.extractInstanceInternalFileShareFileId(uri))
-                } else if (UriUtils.isInstanceInternalFileUrl(user.baseUrl!!, uri)) {
-                    // https://cloud.nextcloud.com/apps/files/?dir=/Engineering&fileid=41
-                    val fileViewerUtils = FileViewerUtils(applicationContext, user)
-                    fileViewerUtils.openFileInFilesApp(uri, UriUtils.extractInstanceInternalFileFileId(uri))
-                } else if (UriUtils.isInstanceInternalFileUrlNew(user.baseUrl!!, uri)) {
-                    // https://cloud.nextcloud.com/apps/files/?dir=/Engineering&fileid=41
-                    val fileViewerUtils = FileViewerUtils(applicationContext, user)
-                    fileViewerUtils.openFileInFilesApp(uri, UriUtils.extractInstanceInternalFileFileIdNew(uri))
-                } else {
-                    super.startActivity(intent)
-                }
-            } else {
-                super.startActivity(intent)
-            }
-        } else {
-            super.startActivity(intent)
-        }
     }
 
     fun sendSelectLocalFileIntent() {
